Sprint 9

+++ src/game/entities/entity.ts
@@ -1,7 +1,14 @@
 import { TPosition } from '@game/@types'
 import { ContextController, GameClock } from '@game/controllers'
-import { Projectile, Vector } from '@game/entities'
+import { Projectile, Vector, Sound } from '@game/entities'
 import { explosion } from '@images'
+import {
+  explosionSound1,
+  explosionSound2,
+  fireSound1,
+  fireSound2,
+  fireSound3
+} from '@game/sound'
 
 class Entity {
   readonly image = new Image()
@@ -17,6 +24,7 @@ class Entity {
   firePeriod: number
   fireCooldown: number
   fireQuantity: number
+  fireAngle: number
   protected projectiles: Array<Projectile>
   protected fire: (context: ContextController) => void
   deathAnimation: () => void
@@ -47,6 +55,7 @@ class Entity {
     this.firePeriod = 50
     this.fireCooldown = this.firePeriod
     this.fireQuantity = 1
+    this.fireAngle = 1
     this.projectileVelocity = 10
     this.velocity = new Vector(velocity)
     this.fire = () => {}
@@ -81,52 +90,102 @@ class Entity {
     clock: GameClock,
     src: string,
     options?: {
-      player?: boolean
       spread?: number
       positionOffset?: TPosition
+      maxLimit?: number
+      callbackWave?: (projectiles: Array<Projectile>) => void
+      callbackEach?: (projectile: Projectile) => void
+      soundURL?: Array<string>
     }
   ) => {
-    const { player, spread, positionOffset } = options ?? {}
+    const { spread, positionOffset, maxLimit, soundURL: tempSoundURL } =
+      options ?? {}
+    const soundURL = tempSoundURL ?? [fireSound1, fireSound2, fireSound3]
+    const sounds: Array<Sound> = []
+    soundURL?.forEach((sound) => {
+      sounds.push(new Sound(sound, { initialVolume: 0.1 }))
+    })
     return (context: ContextController) => {
       this.fireCooldown--
       if (this.fireCooldown <= 0) {
         this.fireCooldown = this.firePeriod
+        if (sounds.length !== 0) {
+          const randomSoundIndex = Math.floor(Math.random() * sounds.length)
+          sounds[randomSoundIndex].stop()
+          sounds[randomSoundIndex].play(context.soundVolume)
+        }
         for (let i = 0; i < this.fireQuantity; i++) {
-          if (this.projectiles.length >= 200) {
+          if (this.projectiles.length >= (maxLimit ?? 200)) {
             this.projectiles[0].deathAnimation = () => {}
-            this.projectiles[0].kill()
+            this.projectiles[0].delete()
             this.projectiles.shift()
           }
+
           const projectile = new Projectile(
             this.projectileVelocity,
             10,
             src,
             this.damage
           )
+
           this.projectiles.push(projectile)
+
           const projectileSpread = spread ?? 0.5
+
           const angle =
             ((i - (this.fireQuantity - 1) / 2) * projectileSpread) /
             this.fireQuantity
+
           const position = positionOffset
             ? {
                 x: this.position.x + positionOffset.x,
                 y: this.position.y + positionOffset.y
               }
             : this.position
-          projectile.init(clock, context, position, player ? angle : angle - 1)
+
+          projectile.init(clock, context, position, angle + this.fireAngle)
+
+          const callback = options?.callbackEach
+            ? options.callbackEach
+            : () => {}
+
+          callback(projectile)
         }
+
+        const callback = options?.callbackWave ? options.callbackWave : () => {}
+
+        callback(
+          this.projectiles.slice(
+            Math.max(this.projectiles.length - this.fireQuantity, 0)
+          )
+        )
       }
     }
   }
 
-  protected initDeathAnimation = (clock: GameClock, src = explosion) => {
+  protected initDeathAnimation = (
+    clock: GameClock,
+    src = explosion,
+    options?: {
+      soundURL?: Array<string>
+    }
+  ) => {
     const image = new Image()
     image.src = src
+    const sounds: Array<Sound> = []
+    const soundURL = options?.soundURL ?? [explosionSound1, explosionSound2]
+    soundURL.forEach((sound) => {
+      sounds.push(new Sound(sound))
+    })
     return () => {
       let size = this.size * 2
       let opacity = 1
       const deathAnimation = clock.startEvent((context) => {
+        if (opacity === 1 && sounds.length !== 0) {
+          const randomSoundIndex = Math.floor(Math.random() * sounds.length)
+          sounds[randomSoundIndex].stop()
+          sounds[randomSoundIndex].play(context.soundVolume)
+        }
         size -= 1
         opacity = opacity - 0.03
         if (opacity <= 0) {
@@ -145,8 +204,8 @@ class Entity {
     const { width, height } = context.getSize()
     const { ox, oy } = context.center
     this.moveTo(
-      Math.random() * (width / 2) + width / 2 - ox,
-      Math.random() * height - oy
+      Math.random() * (width / 2 - this.size) + width / 2 + this.size / 2 - ox,
+      Math.random() * (height - this.size) + this.size / 2 - oy
     )
   }
 
@@ -170,15 +229,38 @@ class Entity {
   }
 
   protected getBoundByBorders = (
-    top: number,
-    right: number,
-    bottom: number,
-    left: number
+    borders: { top: number; right: number; bottom: number; left: number },
+    sizeX: number = 0,
+    sizeY?: number
   ) => {
+    const { top, right, bottom, left } = borders
     const { x, y } = this.position
-    if (x > right || x < left || y > bottom || y < top) {
-      this.position.x = x > right ? right : x < left ? left : this.position.x
-      this.position.y = y > bottom ? bottom : y < top ? top : this.position.y
+    const offsetX = sizeX / 2
+    const offsetY = sizeY ?? sizeX / 2
+    const [offsetRight, offsetLeft, offsetBottom, offsetTop] = [
+      right - offsetX,
+      left + offsetX,
+      bottom - offsetY,
+      top + offsetY
+    ]
+    if (
+      x > offsetRight ||
+      x < offsetLeft ||
+      y > offsetBottom ||
+      y < offsetTop
+    ) {
+      this.position.x =
+        x > offsetRight
+          ? offsetRight
+          : x < offsetLeft
+          ? offsetLeft
+          : this.position.x
+      this.position.y =
+        y > offsetBottom
+          ? offsetBottom
+          : y < offsetTop
+          ? offsetTop
+          : this.position.y
       return true
     }
   }
@@ -200,21 +282,29 @@ class Entity {
     }
   }
 
-  public takeDamage = (damage: number, dispatcher: (score: number) => void) => {
+  public takeDamage = (damage: number) => {
     this.health -= damage
-    if (this.health <= 0) this.kill(dispatcher)
+    if (this.health <= 0) this.kill()
   }
 
   public hit = () => {
     this.kill()
   }
 
-  public kill = (dispatcher: (score: number) => void = () => {}) => {
+  public kill = () => {
+    if (!this.isAlive) return
     this.projectiles.forEach((projectile) => projectile.kill())
     this.isAlive = false
     this.clockEvent()
     this.deathAnimation()
-    dispatcher(this.reward.score)
+    this.killCallback()
+  }
+
+  public delete = () => {
+    if (!this.isAlive) return
+    this.projectiles.forEach((projectile) => projectile.delete())
+    this.isAlive = false
+    this.clockEvent()
     this.killCallback()
   }
 }
